Правка по результатам Code Review
https://github.com/k0st1a/gophermart/pull/1#discussion_r1621327308
